Update SendMessage and Set/GetWindowsLong (#7586)
This changes to a correctly typed WPARAM (unsigned) and LPARAM (signed).

Adds a number of explicit cast operators to try and do the right thing when passing/extracting values through WPARAM, LPARAM, LRESULT.

+++ src/System.Windows.Forms.Design/src/System/Windows/Forms/Design/Behavior/BehaviorService.cs
@@ -832,7 +832,7 @@ namespace System.Windows.Forms.Design.Behavior
                 return;
             }
 
-            if (m.WParamInternal < text.Length + 1)
+            if ((int)m.WParamInternal < text.Length + 1)
             {
                 m.ResultInternal = (LRESULT)(-1);
                 return;
